Refactor youtube feeds
