+++ apps/website/src/app/feeds/youtube/alveus-sanctuary.xml/route.ts
@@ -1,42 +1,21 @@
-import { fetchYouTubeVideos } from "@/server/apis/youtube";
-
 import { channels as youTubeChannels } from "@/data/youtube";
 
-import { getRssFeedContent } from "@/utils/rss-feed";
+import { getYouTubeRssFeedContent } from "@/utils/rss-feed";
 
 export async function GET() {
   const channelUrl = youTubeChannels.alveus.uri;
-  const channels = [youTubeChannels.alveus.id];
-  const latestVideos = await Promise.all(
-    channels.map((channelId) => fetchYouTubeVideos(channelId)),
-  ).then((feeds) =>
-    feeds.flat().sort((a, b) => b.published.getTime() - a.published.getTime()),
+  const channelIds = [youTubeChannels.alveus.id];
+
+  const videoFeedContent = await getYouTubeRssFeedContent(
+    {
+      title: "Alveus Sanctuary - Main YouTube Channel Videos",
+      description:
+        "A feed for new videos on the main Alveus Sanctuary YouTube channel",
+      id: "alveus:youtube:main-channel",
+      link: channelUrl,
+    },
+    channelIds,
   );
-
-  const latestVideoDate = latestVideos[0]?.published;
-
-  const videoFeedItems = latestVideos
-    .map((video) => ({
-      ...video,
-      url: `https://www.youtube.com/watch?v=${encodeURIComponent(video.id)}`,
-    }))
-    .map((video) => ({
-      title: video.title,
-      id: video.url,
-      link: video.url,
-      description: undefined,
-      date: video.published,
-    }));
-
-  const videoFeedContent = getRssFeedContent({
-    title: "Alveus Sanctuary - Main YouTube Channel Videos",
-    description:
-      "A feed for new videos on the main Alveus Sanctuary YouTube channel",
-    id: "alveus:youtube:main-channel",
-    link: channelUrl,
-    updated: latestVideoDate,
-    items: videoFeedItems,
-  });
 
   return new Response(videoFeedContent, {
     headers: {
